refactor: move components to separate folders
