+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Container, Flex } from "@chakra-ui/react";
-import Editor from "./components/Editor/Editor";
+import EditorLayout from "./components/EditorLayout/EditorLayout";
 import Navbar from "./components/Navbar/Navbar";
 
 const App = () => {
@@ -7,7 +7,7 @@ const App = () => {
     <Container px={0} maxW="container.xlg" h="100vh">
       <Flex h="100%" direction="column">
         <Navbar />
-        <Editor />
+        <EditorLayout />
       </Flex>
     </Container>
   );
