+++ src/components/Sidebar/Sidebar.tsx
@@ -1,0 +1,3 @@
+export default () => {
+  return <div>Hello Sidebar</div>
+}
